refactor: 과도한 클라이언트 컴포넌트 분리 코드 제거
