+++ __test__/EventModal.test.tsx
@@ -1,5 +1,4 @@
-import type { ComponentType } from 'react';
-import { act, render, screen, waitFor } from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
 import { describe, expect, it, vi } from 'vitest';
 
 const storage = vi.hoisted(() => ({ get: vi.fn(), set: vi.fn() }));
@@ -12,40 +11,21 @@ vi.mock('@/src/shared/ui/AnimateModal', () => ({
 
 import { EventModal } from '@/src/widgets/event/ui/EventModal';
 
-const EventModalWithLocation = EventModal as ComponentType<{ locationKey: string }>;
-
 describe('EventModal', () => {
-  it('reevaluates the dismissal window when its server-derived location key changes', async () => {
+  it('keeps the modal closed while the dismissal window is active', async () => {
     storage.get.mockReturnValue(new Date(Date.now() + 60_000).toISOString());
 
-    const { rerender } = render(<EventModalWithLocation locationKey="" />);
+    render(<EventModal />);
 
     await waitFor(() => {
       expect(screen.getByTestId('event-modal')).toHaveAttribute('data-open', 'false');
-    });
-
-    storage.get.mockReturnValue(null);
-    rerender(<EventModalWithLocation locationKey="source=header" />);
-
-    await waitFor(() => {
-      expect(screen.getByTestId('event-modal')).toHaveAttribute('data-open', 'true');
     });
   });
 
-  it.each([
-    ['the OAuth location-change event', () => window.dispatchEvent(new Event('goldhand:locationchange'))],
-    ['browser back or forward navigation', () => window.dispatchEvent(new PopStateEvent('popstate'))],
-  ])('reevaluates the dismissal window after %s', async (_reason, notifyLocationChange) => {
-    storage.get.mockReturnValue(new Date(Date.now() + 60_000).toISOString());
-
-    render(<EventModalWithLocation locationKey="" />);
-
-    await waitFor(() => {
-      expect(screen.getByTestId('event-modal')).toHaveAttribute('data-open', 'false');
-    });
-
+  it('opens the modal when no dismissal window is stored', async () => {
     storage.get.mockReturnValue(null);
-    act(notifyLocationChange);
+
+    render(<EventModal />);
 
     await waitFor(() => {
       expect(screen.getByTestId('event-modal')).toHaveAttribute('data-open', 'true');
