--- src/shared/lib/locationChange.ts
@@ -1,5 +1,0 @@
-export const locationChangeEvent = 'goldhand:locationchange';
-
-export const notifyLocationChange = () => {
-  window.dispatchEvent(new Event(locationChangeEvent));
-};
